동영상 목록 bottomBar 모두 선택, 동영상 상세 정보 기능 구현

+++ app/src/main/java/com/dogeby/tagplayer/data/video/source/local/VideoLocalDataSourceImpl.kt
@@ -67,7 +67,7 @@ class VideoLocalDataSourceImpl @Inject constructor(
                             fileName = cursor.getString(nameColumn),
                             duration = cursor.getInt(durationColumn),
                             path = cursor.getString(pathColumn),
-                            size = cursor.getInt(sizeColumn),
+                            size = cursor.getLong(sizeColumn),
                         ),
                     )
                 }
